Pre-render static content
This is so that search engines can see the static content within the
react portions of the site. This doesn't do anything about server side
rendering of dynamic content (e.g. user or project pages). That will
have to be an integration with the backend. But this is a good start.